Now healthbar works, and the player dies when health = 0. Also, each attack has its own strength

+++ Assets/_HOG/Scripts/Interface/HOGHealthBar.cs
@@ -11,7 +11,7 @@ namespace HOG.Screens
     {
         [SerializeField] Slider slider;
 
-        public void SetHealth(int healthPercentage)
+        public void SetHealth(float healthPercentage)
         {
             slider.value = healthPercentage;
         }
